refactor: split worker/baselayout
baselayout is a role suitable for all hosts in our ecosystem.
it installs packages, sets hostname and so on; preparing the
host for a specific task. It can be seen as slightly overzealous
(for instance installing gcc on a backup host), but this can be
solved later with checking specific roles.

also, fix: `::1` in `/etc/hosts/` and start working on linter
as well as jenkins worker.

things are probs broken!